move event submission into standalone file
